Updated incoming webhook implementation as per the review comments
