+++ pkg/adapter/incoming_test.go
@@ -20,10 +20,10 @@ import (
 	"github.com/openshift-pipelines/pipelines-as-code/pkg/provider/gitlab"
 	testclient "github.com/openshift-pipelines/pipelines-as-code/pkg/test/clients"
 	"github.com/openshift-pipelines/pipelines-as-code/pkg/test/kubernetestint"
-	"github.com/openshift-pipelines/pipelines-as-code/pkg/test/logger"
 	"go.uber.org/zap"
 	zapobserver "go.uber.org/zap/zaptest/observer"
 	"gotest.tools/v3/assert"
+	"gotest.tools/v3/env"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	rtesting "knative.dev/pkg/reconciler/testing"
@@ -83,7 +83,8 @@ func Test_compareSecret(t *testing.T) {
 
 func Test_listener_detectIncoming(t *testing.T) {
 	const goodURL = "https://matched/by/incoming"
-	t.Setenv("SYSTEM_NAMESPACE", "pipelinesascode")
+	envRemove := env.PatchAll(t, map[string]string{"SYSTEM_NAMESPACE": "pipelinesascode"})
+	defer envRemove()
 	testNamespace := &corev1.Namespace{
 		ObjectMeta: metav1.ObjectMeta{
 			Name: os.Getenv("SYSTEM_NAMESPACE"),
@@ -609,174 +610,4 @@ func Test_listener_processIncoming(t *testing.T) {
 			assert.Assert(t, l.event.Repository == tt.wantRepo)
 		})
 	}
-}
-
-func Test_GenerateJWT(t *testing.T) {
-	t.Setenv("SYSTEM_NAMESPACE", "foo")
-	namespaceWhereSecretNotInstalled := &corev1.Namespace{
-		ObjectMeta: metav1.ObjectMeta{
-			Name: os.Getenv("SYSTEM_NAMESPACE"),
-		},
-	}
-
-	t.Setenv("SYSTEM_NAMESPACE", "pipelinesascode")
-	testNamespace := &corev1.Namespace{
-		ObjectMeta: metav1.ObjectMeta{
-			Name: os.Getenv("SYSTEM_NAMESPACE"),
-		},
-	}
-
-	validSecret := &corev1.Secret{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      pipelineascode.DefaultPipelinesAscodeSecretName,
-			Namespace: testNamespace.Name,
-		},
-		Data: map[string][]byte{
-			"github-application-id": []byte("12345"),
-			"github-private-key":    []byte(fakePrivateKey),
-		},
-	}
-
-	secretWithInavlidAppID := &corev1.Secret{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      pipelineascode.DefaultPipelinesAscodeSecretName,
-			Namespace: testNamespace.Name,
-		},
-		Data: map[string][]byte{
-			"github-application-id": []byte("abcdf"),
-			"github-private-key":    []byte(fakePrivateKey),
-		},
-	}
-	secretWithInvalidPrivateKey := &corev1.Secret{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      pipelineascode.DefaultPipelinesAscodeSecretName,
-			Namespace: testNamespace.Name,
-		},
-		Data: map[string][]byte{
-			"github-application-id": []byte("12345"),
-			"github-private-key":    []byte("invalidprivatekey"),
-		},
-	}
-
-	tests := []struct {
-		name      string
-		wantErr   bool
-		secrets   []*corev1.Secret
-		namespace []*corev1.Namespace
-	}{{
-		name:      "secret not found",
-		namespace: []*corev1.Namespace{namespaceWhereSecretNotInstalled},
-		secrets:   []*corev1.Secret{},
-		wantErr:   true,
-	}, {
-		name:      "invalid github-application-id",
-		namespace: []*corev1.Namespace{testNamespace},
-		secrets:   []*corev1.Secret{secretWithInavlidAppID},
-		wantErr:   true,
-	}, {
-		name:      "invalid private key",
-		namespace: []*corev1.Namespace{testNamespace},
-		secrets:   []*corev1.Secret{secretWithInvalidPrivateKey},
-		wantErr:   true,
-	}, {
-		name:      "valid secret found",
-		namespace: []*corev1.Namespace{testNamespace},
-		secrets:   []*corev1.Secret{validSecret},
-		wantErr:   false,
-	}}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			logger, _ := logger.GetLogger()
-			tdata := testclient.Data{
-				Namespaces: tt.namespace,
-				Secret:     tt.secrets,
-			}
-			ctxNoSecret, _ := rtesting.SetupFakeContext(t)
-			stdata, _ := testclient.SeedTestData(t, ctxNoSecret, tdata)
-			run := &params.Run{
-				Clients: clients.Clients{
-					Log:            logger,
-					PipelineAsCode: stdata.PipelineAsCode,
-					Kube:           stdata.Kube,
-				},
-			}
-
-			token, err := GenerateJWT(ctxNoSecret, run)
-			if tt.wantErr {
-				assert.Assert(t, err != nil)
-				return
-			}
-			assert.NilError(t, err)
-
-			if token == "" {
-				t.Errorf("there should be a generated token")
-			}
-		})
-	}
-}
-
-func Test_GetAndUpdateInstallationID(t *testing.T) {
-	t.Setenv("SYSTEM_NAMESPACE", "pipelinesascode")
-	testNamespace := &corev1.Namespace{
-		ObjectMeta: metav1.ObjectMeta{
-			Name: os.Getenv("SYSTEM_NAMESPACE"),
-		},
-	}
-
-	validSecret := &corev1.Secret{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      pipelineascode.DefaultPipelinesAscodeSecretName,
-			Namespace: testNamespace.Name,
-		},
-		Data: map[string][]byte{
-			"github-application-id": []byte("274799"),
-			"github-private-key":    []byte(fakePrivateKey),
-		},
-	}
-	logger, _ := logger.GetLogger()
-	tdata := testclient.Data{
-		Namespaces: []*corev1.Namespace{testNamespace},
-		Secret:     []*corev1.Secret{validSecret},
-	}
-	ctx, _ := rtesting.SetupFakeContext(t)
-	stdata, _ := testclient.SeedTestData(t, ctx, tdata)
-	run := &params.Run{
-		Clients: clients.Clients{
-			Log:            logger,
-			PipelineAsCode: stdata.PipelineAsCode,
-			Kube:           stdata.Kube,
-		},
-		Info: info.Info{},
-	}
-
-	l := &listener{
-		run:    run,
-		logger: logger,
-		event:  info.NewEvent(),
-	}
-	req := httptest.NewRequest("GET", "http://localhost", strings.NewReader(""))
-
-	req.Header.Add("X-GitHub-Enterprise-Host", "https://api.github.com")
-
-	repo := &v1alpha1.Repository{
-		ObjectMeta: metav1.ObjectMeta{
-			Name: "repo",
-		},
-		Spec: v1alpha1.RepositorySpec{
-			URL: "https://matched/by/incoming",
-			Incomings: &[]v1alpha1.Incoming{
-				{
-					Targets: []string{"main"},
-					Secret: v1alpha1.Secret{
-						Name: "secret",
-					},
-				},
-			},
-		},
-	}
-
-	err := l.getAndUpdateInstallationID(ctx, req, run, repo)
-	// It always returns error because we don't have valid Private Key to connect with GithubAPP
-	assert.Assert(t, err != nil)
 }
